% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffShrinkHDR.R
\name{diffShrinkHDR}
\alias{diffShrinkHDR}
\title{High Dimensional Regression with Different Shrinkage}
\usage{
diffShrinkHDR(X, Y, Index_XI = NULL, alpha)
}
\arguments{
\item{X}{Design matrix of predictors (n x p)}

\item{Y}{Response vector of length n}

\item{Index_XI}{Optional vector of indices for parameters of interest. If NULL, they will be selected automatically}

\item{alpha}{Significance level for confidence intervals and hypothesis tests (e.g., 0.05)}
}
\description{
High Dimensional Regression with Different Shrinkage
}
